+++ src/__internal__/sticky-footer/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./sticky-footer.component";
-export type { StickyFooterProps } from "./sticky-footer.component";